feat: moving the menu around
