+++ packages/api-client/src/views/Request/RequestSidebarItem.vue
@@ -1,36 +1,24 @@
 <script setup lang="ts">
 import { HttpMethod } from '@/components/HttpMethod'
-import DeleteSidebarListElement from '@/components/Sidebar/Actions/DeleteSidebarListElement.vue'
-import RenameSidebarListElement from '@/components/Sidebar/Actions/RenameSidebarListElement.vue'
 import { useSidebar } from '@/hooks'
 import { getModifiers } from '@/libs'
 import { commandPaletteBus } from '@/libs/event-busses'
 import { PathId } from '@/router'
 import { useWorkspace } from '@/store'
 import {
-  ScalarButton,
-  ScalarContextMenu,
-  ScalarIcon,
-  ScalarModal,
-  useModal,
-} from '@scalar/components'
+  type Item,
+  requestSidebarMenuBus,
+} from '@/views/Request/libs/event-busses'
+import { ScalarButton, ScalarIcon } from '@scalar/components'
 import {
   Draggable,
   type DraggableProps,
   type DraggingItem,
   type HoveredItem,
 } from '@scalar/draggable'
-import type {
-  Collection,
-  Request,
-  RequestExample,
-  RequestMethod,
-  Tag,
-} from '@scalar/oas-utils/entities/spec'
+import type { Request } from '@scalar/oas-utils/entities/spec'
 import { computed, ref } from 'vue'
-import { RouterLink, useRouter } from 'vue-router'
-
-import RequestSidebarItemMenu from './RequestSidebarItemMenu.vue'
+import { RouterLink } from 'vue-router'
 
 const props = withDefaults(
   defineProps<{
@@ -78,20 +66,7 @@ const {
   requestExampleMutators,
   router,
 } = useWorkspace()
-const { replace } = useRouter()
 const { collapsedSidebarFolders, toggleSidebarFolder } = useSidebar()
-
-type Item = {
-  title: string
-  entity: Collection | Tag | Request | RequestExample
-  resourceTitle: string
-  children: string[]
-  method?: RequestMethod
-  link?: string
-  warning?: string
-  rename: () => void
-  delete: () => void
-}
 
 /** Normalize properties across different types for easy consumption */
 const item = computed<Item>(() => {
@@ -108,8 +83,8 @@ const item = computed<Item>(() => {
       children: collection.children,
       warning:
         'This cannot be undone. You’re about to delete the collection and all folders andrequests inside it.',
-      rename: () =>
-        collectionMutators.edit(collection.uid, 'info.title', tempName.value),
+      rename: (name: string) =>
+        collectionMutators.edit(collection.uid, 'info.title', name),
       delete: () =>
         collectionMutators.delete(collection, activeWorkspace.value),
     }
@@ -122,7 +97,7 @@ const item = computed<Item>(() => {
       children: tag.children,
       warning:
         'This cannot be undone. You’re about to delete the tag and all requests inside it',
-      rename: () => tagMutators.edit(tag.uid, 'name', tempName.value),
+      rename: (name: string) => tagMutators.edit(tag.uid, 'name', name),
       delete: () => tagMutators.delete(tag, props.parentUids[0]),
     }
 
@@ -135,8 +110,8 @@ const item = computed<Item>(() => {
       resourceTitle: 'Request',
       warning: 'This cannot be undone. You’re about to delete the request.',
       children: request.examples,
-      rename: () =>
-        requestMutators.edit(request.uid, 'summary', tempName.value),
+      rename: (name: string) =>
+        requestMutators.edit(request.uid, 'summary', name),
       delete: () => requestMutators.delete(request, props.parentUids[0]),
     }
 
@@ -147,8 +122,8 @@ const item = computed<Item>(() => {
     entity: requestExample,
     resourceTitle: 'Example',
     children: [],
-    rename: () =>
-      requestExampleMutators.edit(requestExample.uid, 'name', tempName.value),
+    rename: (name: string) =>
+      requestExampleMutators.edit(requestExample.uid, 'name', name),
     delete: () => requestExampleMutators.delete(requestExample),
   }
 })
@@ -234,32 +209,6 @@ const _isDroppable = (draggingItem: DraggingItem, hoveredItem: HoveredItem) => {
   return true
 }
 
-const tempName = ref('')
-const renameModal = useModal()
-const deleteModal = useModal()
-
-const handleItemRename = (newName: string) => {
-  tempName.value = newName
-  item.value.rename()
-  renameModal.hide()
-}
-
-const openRenameModal = () => {
-  tempName.value = item.value.title || ''
-  renameModal.show()
-}
-
-/** Delete with redirect for both requests and requestExamples */
-const handleItemDelete = () => {
-  item.value.delete()
-
-  if (activeRouterParams.value[PathId.Request] === props.uid)
-    replace(`/workspace/${activeWorkspace.value.uid}/request/default`)
-
-  if (activeRouterParams.value[PathId.Examples] === props.uid)
-    replace(`/workspace/${activeWorkspace.value}/request/default`)
-}
-
 const handleNavigation = (event: KeyboardEvent, _item: Item) => {
   if (event) {
     const modifier = getModifiers(['default'])
@@ -308,120 +257,112 @@ function openCommandPaletteRequest() {
         @click.prevent="
           (event: KeyboardEvent) => handleNavigation(event, item)
         ">
-        <ScalarContextMenu :disabled="isReadOnly">
-          <template #trigger>
-            <div
-              class="group relative flex min-h-8 cursor-pointer flex-row items-start justify-between gap-2 py-1.5 pr-2 rounded editable-sidebar-hover w-full"
-              :class="[
-                highlightClasses,
-                isExactActive || isDefaultActive
-                  ? 'bg-sidebar-active-b text-sidebar-active-c transition-none'
-                  : 'text-sidebar-c-2',
-              ]"
-              tabindex="0">
-              <span
-                class="line-clamp-3 z-10 font-medium w-full pl-2 word-break-break-word"
-                :class="{
-                  'editable-sidebar-hover-item': !isReadOnly,
-                }">
-                {{ item.title }}
-              </span>
-              <div class="flex flex-row gap-1 items-center">
-                <!-- Menu -->
-                <div class="relative">
-                  <RequestSidebarItemMenu
-                    v-if="!isReadOnly"
-                    :item="item.entity"
-                    :parentUids="parentUids"
-                    :resourceTitle="item.resourceTitle"
-                    @delete="deleteModal.show()"
-                    @rename="openRenameModal" />
-                </div>
-                <span class="flex items-start">
-                  &hairsp;
-                  <HttpMethod
-                    v-if="item.method"
-                    class="font-bold"
-                    :method="item.method" />
-                </span>
-              </div>
+        <div
+          class="group relative flex min-h-8 cursor-pointer flex-row items-start justify-between gap-2 py-1.5 pr-2 rounded editable-sidebar-hover w-full"
+          :class="[
+            highlightClasses,
+            isExactActive || isDefaultActive
+              ? 'bg-sidebar-active-b text-sidebar-active-c transition-none'
+              : 'text-sidebar-c-2',
+          ]"
+          tabindex="0">
+          <span
+            class="line-clamp-3 z-10 font-medium w-full pl-2 word-break-break-word"
+            :class="{
+              'editable-sidebar-hover-item': !isReadOnly,
+            }">
+            {{ item.title }}
+          </span>
+          <div class="flex flex-row gap-1 items-center">
+            <!-- Menu -->
+            <div class="relative">
+              <ScalarButton
+                v-if="!isReadOnly"
+                class="px-0.5 py-0 z-10 hover:bg-b-3 hidden group-hover:flex ui-open:flex absolute -translate-y-1/2 right-0 aspect-square inset-y-2/4 h-fit"
+                size="sm"
+                type="button"
+                variant="ghost"
+                @click.stop.prevent="
+                  (ev) =>
+                    requestSidebarMenuBus.emit({
+                      item,
+                      parentUids,
+                      targetRef: ev.currentTarget,
+                    })
+                ">
+                <ScalarIcon
+                  icon="Ellipses"
+                  size="sm" />
+              </ScalarButton>
             </div>
-          </template>
-          <template #content>
-            <RequestSidebarItemMenu
-              :item="item.entity"
-              :parentUids="parentUids"
-              :resourceTitle="item.resourceTitle"
-              static
-              @delete="deleteModal.show()"
-              @rename="openRenameModal" />
-          </template>
-        </ScalarContextMenu>
+            <span class="flex items-start">
+              &hairsp;
+              <HttpMethod
+                v-if="item.method"
+                class="font-bold"
+                :method="item.method" />
+            </span>
+          </div>
+        </div>
       </RouterLink>
 
       <!-- Collection/Folder -->
-      <ScalarContextMenu
+      <button
         v-else-if="!isReadOnly || parentUids.length"
-        :disabled="isReadOnly || isDraftCollection">
-        >
-        <template #trigger>
-          <button
-            class="hover:bg-b-2 group relative flex w-full flex-row justify-start gap-1.5 rounded p-1.5 z-[1]"
-            :class="highlightClasses"
-            type="button"
-            @click="toggleSidebarFolder(item.entity.uid)">
-            <span
-              class="z-10 flex h-5 items-center justify-center max-w-[14px]">
-              <slot name="leftIcon">
-                <div
-                  :class="{
-                    'rotate-90': collapsedSidebarFolders[item.entity.uid],
-                  }">
-                  <ScalarIcon
-                    class="text-c-3 text-sm"
-                    icon="ChevronRight"
-                    size="sm"
-                    thickness="2.5" />
-                </div>
-              </slot>
-              &hairsp;
-            </span>
+        class="hover:bg-b-2 group relative flex w-full flex-row justify-start gap-1.5 rounded p-1.5 z-[1]"
+        :class="highlightClasses"
+        :disabled="isReadOnly"
+        type="button"
+        @click="toggleSidebarFolder(item.entity.uid)">
+        <span class="z-10 flex h-5 items-center justify-center max-w-[14px]">
+          <slot name="leftIcon">
             <div
-              class="flex flex-1 flex-row justify-between editable-sidebar-hover">
-              <span
-                class="line-clamp-3 z-10 font-medium text-left w-full word-break-break-word"
-                :class="{
-                  'editable-sidebar-hover-item': !isReadOnly,
-                }">
-                {{ item.title }}
-              </span>
-              <div class="relative flex h-fit">
-                <RequestSidebarItemMenu
-                  v-if="!isReadOnly && !isDraftCollection"
-                  :item="item.entity"
-                  :parentUids="parentUids"
-                  :resourceTitle="item.resourceTitle"
-                  @delete="deleteModal.show()"
-                  @rename="openRenameModal" />
-                <span>&hairsp;</span>
-              </div>
+              :class="{
+                'rotate-90': collapsedSidebarFolders[item.entity.uid],
+              }">
+              <ScalarIcon
+                class="text-c-3 text-sm"
+                icon="ChevronRight"
+                size="sm"
+                thickness="2.5" />
             </div>
-          </button>
-        </template>
-        <template #content>
-          <RequestSidebarItemMenu
-            v-if="!isReadOnly && !isDraftCollection"
-            :item="item.entity"
-            :parentUids="parentUids"
-            :resourceTitle="item.resourceTitle"
-            static
-            @delete="deleteModal.show()"
-            @rename="openRenameModal" />
-        </template>
-      </ScalarContextMenu>
+          </slot>
+          &hairsp;
+        </span>
+        <div
+          class="flex flex-1 flex-row justify-between editable-sidebar-hover">
+          <span
+            class="line-clamp-3 z-10 font-medium text-left w-full word-break-break-word"
+            :class="{
+              'editable-sidebar-hover-item': !isReadOnly,
+            }">
+            {{ item.title }}
+          </span>
+          <div class="relative flex h-fit">
+            <ScalarButton
+              v-if="!isReadOnly && !isDraftCollection"
+              class="px-0.5 py-0 z-10 hover:bg-b-3 hidden group-hover:flex ui-open:flex absolute -translate-y-1/2 right-0 aspect-square inset-y-2/4 h-fit"
+              size="sm"
+              variant="ghost"
+              @click.stop.prevent="
+                (ev) =>
+                  requestSidebarMenuBus.emit({
+                    item,
+                    parentUids,
+                    targetRef: ev.currentTarget,
+                  })
+              ">
+              <ScalarIcon
+                icon="Ellipses"
+                size="sm" />
+            </ScalarButton>
+            <span>&hairsp;</span>
+          </div>
+        </div>
+      </button>
 
       <!-- Children -->
-      <div v-show="showChildren">
+      <div v-if="showChildren">
         <!-- We never want to show the first example -->
         <RequestSidebarItem
           v-for="childUid in item.children"
@@ -447,25 +388,6 @@ function openCommandPaletteRequest() {
       </div>
     </Draggable>
   </div>
-  <ScalarModal
-    :size="'xxs'"
-    :state="deleteModal"
-    :title="`Delete ${item.resourceTitle}`">
-    <DeleteSidebarListElement
-      :variableName="item.title"
-      :warningMessage="item.warning"
-      @close="deleteModal.hide()"
-      @delete="handleItemDelete" />
-  </ScalarModal>
-  <ScalarModal
-    :size="'xxs'"
-    :state="renameModal"
-    :title="`Rename ${item.resourceTitle}`">
-    <RenameSidebarListElement
-      :name="item.title"
-      @close="renameModal.hide()"
-      @rename="handleItemRename" />
-  </ScalarModal>
 </template>
 
 <style>
